requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
django-cors-headers
python-decouple
